+++ saleor/graphql/query_cost_map.py
@@ -143,6 +143,7 @@ COST_MAP = {
     },
     "Attribute": {
         "choices": {"complexity": 1, "multipliers": ["first", "last"]},
+        "values": {"complexity": 1, "multipliers": ["first", "last"]},
         "productTypes": {"complexity": 1, "multipliers": ["first", "last"]},
         "productVariantTypes": {"complexity": 1, "multipliers": ["first", "last"]},
     },